revert(reactions): drop the full emoji picker from the quick reaction popup
Parked at the user's request after trying the build. This backs out f3a71a6e in
full: the shared EmojiPickerPanel, the useReactionPicker composable, the
randomEmoji helper, and the two entry points into the full list.

The quick reaction popup is back to what it was. On desktop the hover card shows
only your quick reactions again, and on mobile the bottom sheet shows the quick
grid and the reaction summary. Quick reaction settings were never touched.
