+++ frontend/src/utils/emoji.ts
@@ -1,6 +1,3 @@
-import { gemoji } from 'gemoji'
-import { getRandomNumber } from '@/utils'
-
 /**
  * A reaction value is either a unicode emoji (e.g. "👍") or a custom emoji
  * stored as an uploaded image URL. Custom emoji values are file paths, so they
@@ -8,9 +5,4 @@ import { getRandomNumber } from '@/utils'
  */
 export function isImageEmoji(value: string): boolean {
   return /^(https?:\/\/|\/)/.test(value)
-}
-
-/** Any unicode emoji, picked at random. Custom emoji are never returned. */
-export function randomEmoji(): string {
-  return gemoji[getRandomNumber(0, gemoji.length - 1)].emoji
 }
